+++ src/components/activation-displays/activation-description.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { FieldDefinition, LabelsList } from '../labels-list/labels-list'
 import { ParsedToken } from '../parsed-description/parsed-description'
 import { ActionRenderer } from './action-renderer'
-import { Activation, SavingThrow } from '@/models/activation'
+import { Activation, SavingThrow } from '@/models/db/activation'
 
 export function ActivationDescription({
   value,
