+++ src/models/db/character-entity.ts
