+++ src/models/db/equipment.tsx
