+++ src/models/db/spell.tsx
