+++ src/models/db/trait.tsx
