+++ src/utils/services/card-factory.tsx
@@ -3,11 +3,11 @@ import EquipmentCard from '@/components/equipment/equipment-card'
 import EquipmentWithVariantsCard from '@/components/equipment/equipment-with-variants-card'
 import SpellCard from '@/components/spells/spell-card'
 import TraitCard from '@/components/traits/trait-card'
-import Condition from '@/models/condition'
-import { EntityModel } from '@/models/entity-model'
-import { Equipment, EquipmentWithVariants } from '@/models/equipment'
-import { Spell } from '@/models/spell'
-import Trait from '@/models/trait'
+import Condition from '@/models/db/condition'
+import { EntityModel } from '@/models/db/entity-model'
+import { Equipment, EquipmentWithVariants } from '@/models/db/equipment'
+import { Spell } from '@/models/db/spell'
+import Trait from '@/models/db/trait'
 import { ReactNode, RefObject } from 'react'
 
 export function CardFactory<T extends EntityModel>({
